refactor map utils & icons
